[BugFix] Fix the problem of updating MV partition meta incorrectly after MV refresh (#39699)
Why I'm doing:
After the refresh of partitioned MV, the partititon meta(lastRefreshTime) may be
incorrect. That is because the partition order in selectedPartitionNames and
partitions in function PartitionBasedMvRefreshProcessor::getSelectedPartitionInfos
are not the same.

What I'm doing:
Sort selectedPartitionNames before the for loop. Since the partitions is obtained
from the SQL and the SQL has "ORDER BY PARTITION_DESCRIPTION" so partitions have
been sorted.